Revert "Feature/pq 222 make a concise view for approved budgets for researchers"
