+++ package.json
@@ -32,7 +32,7 @@
         "next": "^14.0.1",
         "next-auth": "^4.24.4",
         "nodemailer": "^6.9.7",
-        "pnpm": "^8.10.2",
+        "pnpm": "^8.10.0",
         "react": "^18.2.0",
         "react-csv": "^2.2.2",
         "react-dom": "^18.2.0",
@@ -45,7 +45,7 @@
         "@tailwindcss/typography": "^0.5.10",
         "@types/bcryptjs": "^2.4.5",
         "@types/node": "^20.8.10",
-        "@types/react": "^18.2.34",
+        "@types/react": "^18.2.33",
         "@types/react-dom": "^18.2.14",
         "@typescript-eslint/eslint-plugin": "^6.9.1",
         "@typescript-eslint/parser": "^6.9.1",
