--- src/app/(authenticated)/protocols/[id]/@actions/loading.tsx
@@ -1,9 +1,0 @@
-export default function Loading() {
-    return (
-        <div className="flex animate-pulse items-center justify-end gap-2">
-            <div className="h-9 w-20 rounded-md bg-gray-100" />
-            <div className="h-9 w-28 rounded-md bg-gray-100" />
-            <div className="h-9 w-24 rounded-md bg-gray-100" />
-        </div>
-    )
-}
