--- src/app/(authenticated)/protocols/[id]/@evaluators/loading.tsx
@@ -1,5 +1,0 @@
-export default function Loading() {
-    return (
-        <div className="my-1 ml-2 h-32 w-full max-w-3xl animate-pulse rounded-lg bg-gray-100" />
-    )
-}
