--- src/app/(authenticated)/protocols/[id]/@reviews/loading.tsx
@@ -1,8 +1,0 @@
-export default function Loading() {
-    return (
-        <div className="w-full animate-pulse space-y-3 lg:sticky lg:top-0 lg:w-[28rem] lg:pl-2 xl:w-[36rem]">
-            <div className="h-80 rounded-md bg-gray-100"></div>
-            <div className="h-20 rounded-md bg-gray-100"></div>
-        </div>
-    )
-}
