+++ src/app/(authenticated)/protocols/[id]/page.tsx
@@ -1,7 +1,7 @@
 import View from '@protocol/protocol-view-template'
 import { redirect } from 'next/navigation'
 import { findProtocolById } from 'repositories/protocol'
-import { PDF } from 'modules/protocol-pdf'
+import { PDF } from 'modules/pdf'
 
 export default async function Page({ params }: { params: { id: string } }) {
     if (params.id === 'new') redirect('/protocols/new/0')
