+++ src/modules/protocol/protocol-metadata.tsx
@@ -1,68 +1,74 @@
 import { Badge } from '@elements/badge'
-import type { User } from '@prisma/client'
+import type { State, User } from '@prisma/client'
 import ProtocolStatesDictionary from '@utils/dictionaries/ProtocolStatesDictionary'
 import { dateFormatter } from '@utils/formatters'
 import { Calendar, User as UserIcon } from 'tabler-icons-react'
+import { ResearcherReassignation } from './elements/action-buttons/researcher-reassignation'
 import { getAllOwners } from '@repositories/user'
-import { getServerSession } from 'next-auth'
-import { authOptions } from 'app/api/auth/[...nextauth]/route'
-import { getProtocolMetadata } from '@repositories/protocol'
-import { ResearcherReassignation } from '@protocol/elements/action-buttons/researcher-reassignation'
-import ProtocolNumberUpdate from '@protocol/elements/protocol-number-update'
+import ProtocolNumberUpdate from './elements/protocol-number-update'
 
-export default async function ProtocolMetadata({
-    params,
+export async function ProtocolMetadata({
+    currentUser,
+    id,
+    protocolNumber,
+    createdAt,
+    state,
+    researcher,
+    convocatory,
 }: {
-    params: { id: string }
+    currentUser: User
+    id: string
+    protocolNumber: string | null
+    createdAt: Date | null
+    state: State
+    researcher: { id: string; name: string; email: string }
+    convocatory: { id: string; name: string }
 }) {
-    const session = await getServerSession(authOptions)
-    const protocol = await getProtocolMetadata(params.id)
-    if (!session || !protocol) return
     let researcherList: User[] = []
-    if (session.user.role === 'ADMIN') {
+    if (currentUser.role === 'ADMIN') {
         researcherList = await getAllOwners()
     }
 
     return (
-        <div className="relative z-20 my-1 ml-2 max-w-3xl flex-grow gap-2 rounded-lg bg-gray-50/50 px-3 py-2 leading-relaxed drop-shadow-sm">
+        <div className="z-10 my-1 ml-2 max-w-3xl flex-grow gap-2 rounded-lg bg-gray-50/50 px-3 py-2 leading-relaxed drop-shadow-sm">
             <div className="flex items-center justify-between gap-4">
                 <div className="flex items-center gap-6">
                     <ProtocolNumberUpdate
-                        protocolId={params.id}
-                        protocolNumber={protocol.protocolNumber}
-                        role={session.user.role}
+                        protocolId={id}
+                        protocolNumber={protocolNumber}
+                        role={currentUser.role}
                     />
 
                     <span>
                         <Calendar className="mr-2 inline h-4 text-gray-600" />
                         <span className="mt-0.5 text-xs text-gray-600">
-                            {dateFormatter.format(protocol.createdAt!)}
+                            {dateFormatter.format(createdAt!)}
                         </span>
                     </span>
                 </div>
-                <Badge>{protocol.convocatory.name}</Badge>
+                <Badge>{convocatory?.name}</Badge>
             </div>
             <div className="mt-2 flex items-baseline gap-4">
                 <div className="flex items-center gap-2">
                     <UserIcon className="h-4 text-gray-600" />
                     <div className="font-medium">
-                        {protocol.researcher.name}
+                        {researcher?.name}
                         <div className="-mt-1.5 ml-px text-xs font-light text-gray-500">
-                            {protocol.researcher.email}
+                            {researcher?.email}
                         </div>
                     </div>
                 </div>
                 <div className="mx-3 mt-1 flex-grow">
-                    {session.user.role === 'ADMIN' && (
+                    {currentUser.role === 'ADMIN' && (
                         <ResearcherReassignation
-                            protocolId={params.id}
-                            researcherId={protocol.researcher.id}
+                            protocolId={id}
+                            researcherId={researcher?.id}
                             researchers={researcherList}
                         />
                     )}
                 </div>
                 <Badge className="text-sm">
-                    {ProtocolStatesDictionary[protocol.state]}
+                    {ProtocolStatesDictionary[state]}
                 </Badge>
             </div>
         </div>
